+++ Test Source/TetsInterfacePrinter/Test_Print_Print.java
@@ -1,0 +1,21 @@
+package TetsInterfacePrinter;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class Test_Print_Print {
+    @ParameterizedTest
+    @CsvSource({"testen, "})
+    public void testPrint(String text, String expected){
+        //Arrange
+        TestClassPrinterInterface printer = new TestClassPrinterInterface();
+
+        //Act
+        printer.print(text);
+
+        //Assert
+        assertEquals(expected, TestClassPrinterInterface.outCome);
+    }
+}
